functions/comment-notifications/requirements.txt
azure-functions
azure-communication-email
azure-identity
requests
